requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
python-dotenv
streamlit
streamlit-chat
ipykernel
duckduckgo-search
langgraph